LRF RDO: take rate into account, use SSE modeling
This modifies sgrproj_solve to break with the lkibaom solution
technique.  It uses the model libaom builds, but rather than using it
to minimize error to solve for coefficients, it uses the model to
predict error for a given coefficient set while taking rate into
account to minimize RDO, not just distortion.

Right now it is an exhaustive search (though fast), and does not also
take cdef_dist or distortion bias into account.  All three need to be
corrected for produciton.